BE: Refactored entry point by separating concerns
