+++ server/api/main.py
@@ -1,147 +1,28 @@
-import json
-import os
-
-from datetime import datetime, timedelta
-from typing import Optional
-
-from passlib.context import CryptContext
-from jose import JWTError, jwt
-
-from fastapi import Depends, FastAPI, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from sqlalchemy.orm import Session
-
-from db import crud, models, schemas
-from db.session import get_db, engine
-
-
-with open(os.path.join("secrets.json")) as secrets_file:
-    secrets = json.load(secrets_file)
-
-
-def get_secret(setting, secrets=secrets):
-    try:
-        return secrets[setting]
-    except KeyError:
-        raise KeyError("Improperly Configured")
-
-
-SECRET_KEY = get_secret("SECRET_KEY")
-ALGORITHM = get_secret("ALGORITHM")
-ACCESS_TOKEN_EXPIRE_MINUTES = get_secret("ACCESS_TOKEN_EXPIRE_MINUTES")
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-# temporary test for User login OAuth2 flow
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+from fastapi import Depends, FastAPI
+from api.api_v1.routers.auth import auth_router
+from api.api_v1.routers.users import users_router
+from core.auth import get_current_user
+import uvicorn
 
 # temporary mechanism for migrations
-models.Base.metadata.create_all(bind=engine)
+# models.Base.metadata.create_all(bind=engine)
 
-app = FastAPI()
-
-
-def verify_password(plain_password, hashed_password):
-    return pwd_context.verify(plain_password, hashed_password)
+app = FastAPI(docs_url="/api/docs", openapi_url="/api")
 
 
-def get_password_hash(password):
-    return pwd_context.hash(password)
+@app.get("/api/v1")
+async def root():
+    return {"message": "Hellow Root"}
 
 
-def authenticate_user(username: str, password: str, db: Session = Depends(get_db)):
-    user = crud.get_user_by_username(db=db, username=username)
-    if not user:
-        return False
-    if not verify_password(password, user.hashed_password):
-        return False
-    return user
+# Routers
+app.include_router(
+    users_router,
+    prefix="/api/v1",
+    tags=["users"],
+    dependencies=[Depends(get_current_user)],
+)
+app.include_router(auth_router, prefix="/api", tags=["auth"])
 
-
-def get_user(db: Session, username: str):
-    if username in db:
-        user_dict = db[username]
-        return schemas.UserInDB(**user_dict)
-
-
-def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-
-async def get_current_user(
-    db: Session = Depends(get_db), token: str = Depends(oauth2_scheme)
-):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = schemas.TokenData(username=username)
-    except JWTError:
-        raise credentials_exception
-    user = crud.get_user_by_username(db, username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
-
-@app.post("/token", response_model=schemas.Token)
-async def login_for_access_token(
-    db: Session = Depends(get_db), form_data: OAuth2PasswordRequestForm = Depends()
-):
-    user = authenticate_user(form_data.username, form_data.password, db)
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Incorrect username or password",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    access_token = create_access_token(
-        data={"sub": user.username}, expires_delta=access_token_expires
-    )
-    return {"access_token": access_token, "token_type": "bearer"}
-
-
-@app.get("/users/me", response_model=schemas.User)
-async def read_users_me(current_user=Depends(get_current_user)):
-    return current_user
-
-
-@app.get("/items/")
-async def read_items(token: str = Depends(oauth2_scheme)):
-    return {"token": token}
-
-
-@app.post("/users/", response_model=schemas.User)
-def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
-    db_user = crud.get_user_by_username(db, username=user.username)
-    if db_user:
-        raise HTTPException(status_code=400, details="Username already registered")
-    return crud.create_user(db=db, user=user)
-
-
-@app.get("/users/", response_model=list[schemas.User])
-def read_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-    users = crud.get_users(db, skip=skip, limit=limit)
-    return users
-
-
-@app.get("/users/{user_id}", response_model=schemas.User)
-def read_user(user_id: int, db: Session = Depends(get_db)):
-    db_user = crud.get_user(db, user_id=user_id)
-    if db_user is None:
-        raise HTTPException(status_code=404, detail="User not found")
-    return db_user
+if __name__ == "__main__":
+    uvicorn.run("main:app", host="0.0.0.0", reload=True, port=8000)
